+++ src/test/java/nextstep/subway/line/LineStationAcceptanceTest.java
@@ -7,7 +7,7 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.domain.line.LineStation;
 import nextstep.subway.domain.line.LineStations;
-import nextstep.subway.domain.Station;
+import nextstep.subway.domain.station.Station;
 import nextstep.subway.station.StationAcceptanceTest;
 import org.springframework.http.HttpStatus;
 import nextstep.subway.TestUtil;
