+++ src/test/java/nextstep/subway/line/LineStationAcceptanceTestFixture.java
@@ -6,7 +6,7 @@ import io.restassured.response.Response;
 import java.util.HashMap;
 import java.util.Map;
 import nextstep.subway.domain.line.Line;
-import nextstep.subway.domain.Station;
+import nextstep.subway.domain.station.Station;
 import nextstep.subway.station.StationAcceptanceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.web.server.LocalServerPort;
